Add files via upload
Using Ralph Kimball methodolgy, created a multi-entity dimensional model diagram and database for Taskrabbit and loaded it with data. Then utilized both Qlikview and Tableau each to erase loops and visualize the data into sheets and dashboards highlighting the customer, taskers and task trends in the form of quantity, frequency, distribution, location, etc.